PaymentController retrievePaymentsWithApportion responsetype changed based on IAC changes and changed the test cases
